+++ main-server/src/main/java/ru/practicum/event/controller/UsersEventController.java
@@ -32,7 +32,7 @@ public class UsersEventController {
     private final EventService eventService;
 
     @GetMapping("/{userId}/events")
-    public List<Event> getEventsByUserId(@PathVariable @NotNull Integer userId,
+    public List<Event> getEventsByUserId(@PathVariable @NotNull Long userId,
             @RequestParam(name = "from", defaultValue = "0") @PositiveOrZero int from,
             @RequestParam(name = "size", defaultValue = "10") @Positive int size) {
         log.info("Get events by user id request.");
@@ -41,22 +41,22 @@ public class UsersEventController {
 
     @PostMapping("/{userId}/events")
     @ResponseStatus(HttpStatus.CREATED)
-    public EventFullDto addEvent(@PathVariable @NotNull Integer userId,
+    public EventFullDto addEvent(@PathVariable @NotNull Long userId,
             @RequestBody @NotNull @Valid NewEventDto newEventDto) {
         log.info("Add event request.");
         return eventService.addEvent(userId, newEventDto);
     }
 
     @GetMapping("/{userId}/events/{eventId}")
-    public EventFullDto getEvent(@PathVariable @NotNull Integer userId,
-            @PathVariable @NotNull Integer eventId) {
+    public EventFullDto getEvent(@PathVariable @NotNull Long userId,
+            @PathVariable @NotNull Long eventId) {
         log.info("Get event by event id request.");
         return eventService.getEvent(userId, eventId);
     }
 
     @PatchMapping("/{userId}/events/{eventId}")
-    public EventFullDto updateEvent(@PathVariable @NotNull Integer userId,
-            @PathVariable @NotNull Integer eventId, @RequestBody @NotNull @Valid
+    public EventFullDto updateEvent(@PathVariable @NotNull Long userId,
+            @PathVariable @NotNull Long eventId, @RequestBody @NotNull @Valid
     UpdateEventUserRequest updateEventUserRequest) {
         log.info("Update event request.");
         return eventService.updateEvent(userId, eventId, updateEventUserRequest);
