+++ main-server/src/main/java/ru/practicum/user/repository/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ru.practicum.user.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-    List<User> findAllByIdIn(List<Integer> ids, Pageable pageable);
+    List<User> findAllByIdIn(List<Long> ids, Pageable pageable);
 
     User findByName(String name);
